requirements.dev.txt
asttokens==2.4.1
cachetools==5.0.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
cryptography==42.0.8
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
flake8==7.1.0
fonttools==4.53.1
google-auth==2.31.0
google-auth-oauthlib==1.2.0
greenlet==3.0.3
gspread==6.0.1
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
jsonpath-ng==1.6.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
lxml==5.2.2
matplotlib==3.5.0
matplotlib-inline==0.1.7
mccabe==0.7.0
mypy==1.10.1
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.25.2
oauthlib==3.2.2
openpyxl==3.1.2
packaging==24.1
pandas==2.1.0
parso==0.8.4
pdfminer.six==20231228
pdfplumber==0.11.2
pillow==10.4.0
platformdirs==4.2.2
ply==3.11
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg2==2.9.2
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.0
pycparser==2.22
pydantic==1.10.13
pyflakes==3.2.0
Pygments==2.18.0
pyparsing==3.1.2
pypdfium2==4.30.0
python-dateutil==2.8.2
pytz==2021.3
pywin32==302
PyYAML==6.0.1
pyzmq==26.0.3
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
setuptools-scm==8.1.0
six==1.16.0
SQLAlchemy==2.0.31
stack-data==0.6.3
StrEnum==0.4.15
tenacity==8.0.1
tomli==2.0.1
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2022.2
urllib3==2.2.2
wcwidth==0.2.13
xlrd==2.0.1
xlwings==0.30.11
xmltodict==0.13.0
